+++ crest/Assets/Crest/Crest/Scripts/Interaction/SphereWaterInteraction.cs
@@ -45,6 +45,8 @@ namespace Crest
         [SerializeField]
         bool _warnOnSpeedClamp = false;
 
+        Vector3 _velocity;
+        Vector3 _velocityClamped;
         Vector3 _posLast;
 
         float _weightThisFrame;
@@ -100,11 +102,17 @@ namespace Crest
             _sampleHeightHelper.Init(transform.position, 2f * _radius);
             _sampleHeightHelper.Sample(out Vector3 disp, out _, out _);
 
-            // Enforce upwards
-            transform.rotation = Quaternion.Euler(90f, 0f, 0f);
+            LateUpdateComputeVel(ocean);
 
             // Velocity relative to water
-            Vector3 relativeVelocity = LateUpdateComputeVelRelativeToWater(ocean);
+            var relativeVelocity = _velocityClamped;
+            {
+                _sampleFlowHelper.Init(transform.position, 2f * _radius);
+                _sampleFlowHelper.Sample(out var surfaceFlow);
+                relativeVelocity -= new Vector3(surfaceFlow.x, 0, surfaceFlow.y);
+
+                relativeVelocity.y *= _weightUpDownMul;
+            }
 
             var dt = 1f / ocean._lodDataDynWaves.Settings._simulationFrequency;
 
@@ -137,47 +145,43 @@ namespace Crest
         }
 
         // Velocity of the sphere, relative to the water. Computes on the fly, discards if teleport detected.
-        Vector3 LateUpdateComputeVelRelativeToWater(OceanRenderer ocean)
+        void LateUpdateComputeVel(OceanRenderer ocean)
         {
-            Vector3 vel;
-
-            // feed in water velocity
-            vel = (transform.position - _posLast) / ocean.DeltaTimeDynamics;
+            // Compue vel using finite difference
+            _velocity = (transform.position - _posLast) / ocean.DeltaTimeDynamics;
             if (ocean.DeltaTimeDynamics < 0.0001f)
             {
-                vel = Vector3.zero;
+                _velocity = Vector3.zero;
             }
 
-            {
-                _sampleFlowHelper.Init(transform.position, 2f * _radius);
-                _sampleFlowHelper.Sample(out var surfaceFlow);
-                vel -= new Vector3(surfaceFlow.x, 0, surfaceFlow.y);
-            }
-            vel.y *= _weightUpDownMul;
-
-            var speedKmh = vel.magnitude * 3.6f;
+            var speedKmh = _velocity.magnitude * 3.6f;
             if (speedKmh > _teleportSpeed)
             {
                 // teleport detected
-                vel *= 0f;
+                _velocity *= 0f;
 
                 if (_warnOnTeleport)
                 {
                     Debug.LogWarning("Crest: Teleport detected (speed = " + speedKmh.ToString() + "), velocity discarded.", this);
                 }
+
+                speedKmh = _velocity.magnitude * 3.6f;
             }
-            else if (speedKmh > _maxSpeed)
+
+            if (speedKmh > _maxSpeed)
             {
                 // limit speed to max
-                vel *= _maxSpeed / speedKmh;
+                _velocityClamped = _velocity * _maxSpeed / speedKmh;
 
                 if (_warnOnSpeedClamp)
                 {
                     Debug.LogWarning("Crest: Speed (" + speedKmh.ToString() + ") exceeded max limited, clamped.", this);
                 }
             }
-
-            return vel;
+            else
+            {
+                _velocityClamped = _velocity;
+            }
         }
 
         // Weight based on submerged-amount of sphere
@@ -219,10 +223,25 @@ namespace Crest
             Gizmos.DrawWireSphere(transform.position, _radius);
         }
 
-        public void Draw(CommandBuffer buf, float weight, int isTransition, int lodIdx)
+        public void Draw(LodDataMgr lodData, CommandBuffer buf, float weight, int isTransition, int lodIdx)
         {
+            var timeBeforeCurrentTime = (lodData as LodDataMgrDynWaves).TimeLeftToSimulate;
+
+            // Draw little red markers for each substep position
+            //var pos = transform.position + -_velocity * timeBeforeCurrentTime;
+            //Debug.DrawLine(pos - transform.right + transform.up, pos + transform.right + transform.up, Color.red, 0.5f);
+
+            // _renderMatrix is only updated at the frame update rate, whereas this input wants to apply
+            // to substeps. Reconstruct the position of this input at the current substep time. This produces
+            // much smoother interaction shapes for moving objects. Increasing sim freq helps further.
+            var renderMatrix = _renderMatrix;
+            var offset = _velocity * timeBeforeCurrentTime;
+            renderMatrix.m03 -= offset.x;
+            renderMatrix.m13 -= offset.y;
+            renderMatrix.m23 -= offset.z;
+
             _mpb.SetFloat(sp_weight, weight * _weightThisFrame);
-            buf.DrawMesh(RegisterLodDataInputBase.QuadMesh, _renderMatrix, _mat, 0, 0, _mpb);
+            buf.DrawMesh(RegisterLodDataInputBase.QuadMesh, renderMatrix, _mat, 0, 0, _mpb);
         }
     }