+++ docs/about/history.rst
@@ -43,6 +43,7 @@ Changed
    -  Double sample count for *ShapeGerstner* waves to improve quality.
    -  Tidy up wave spectrum inspector by only showing *ShapeGerstner*-specific controls when editing within a *ShapeGerstner* component.
    -  Add option (enabled by default) to prewarm foam simulation on load and camera teleports.
+   -  *Underwater Renderer* validates *Ocean Renderer* material.
 
 Fixed
 ^^^^^
@@ -64,6 +65,10 @@ Fixed
    -  Fix *Allow Null Light* option on *Sim Settings Shadows* not working.
    -  Fix ocean tiles not reverting to *Ocean Renderer > Material* if *Water Body > Override Material* was used and *Water Body* was disabled or removed.
    -  Add *Time Scale* control for FFT (*Gravity* setting was broken).
+   -  Fix underwater rendering when the camera's culling mask excludes the *Ocean Renderer > Layer*.
+   -  Fix visible "rings" in dynamic wave sim resulting from fast moving objects that have the *Sphere Water Interaction* component attached. Simulation frequency can be increased to improve result further, at the cost of more simulation steps per frame.
+   -  Fix *Sphere Water Interaction* component not working in standalone builds.
+
 
    .. only:: birp